acpi

+++ include/io.h
@@ -23,7 +23,7 @@ static inline uint8_t inb(uint16_t port)
                    : "Nd"(port) );
     return ret;
 }
-static inline uint8_t inw(uint16_t port)
+static inline uint16_t inw(uint16_t port)
 {
     uint16_t ret;
     asm volatile ( "inw %1, %0"
